OCR Enhancement
The OCR has been changed in a few major ways. First, the OCR scans the image itself and then works on the scanned image making it more accurate. Secondly, the background color used for each bounding box is calculated by using the area around the box, and thirdly, the text color takes the 3 most common colors in the bounding box, and sets the text color to the one which has the most contrast against the background color.